apps/brain-service/requirements.txt
# =============================================================================
# CORE FRAMEWORK - FastAPI Foundation
# =============================================================================
fastapi==0.115.0          # Web framework (your proven pattern)
uvicorn[standard]==0.32.0  # ASGI server with performance extras
pydantic==2.10.3           # Data validation and settings

# =============================================================================
# AI PROVIDER CLIENTS - Multi-Model Support  
# =============================================================================
anthropic==0.40.0          # Claude API (primary model)
openai==1.58.1             # OpenAI GPT API 
google-generativeai==0.8.3 # Gemini API
tiktoken==0.8.0            # Token counting for OpenAI

# =============================================================================
# gRPC COMMUNICATION - Identra Integration
# =============================================================================
grpcio==1.69.0             # gRPC core
grpcio-tools==1.69.0       # Protocol buffer compiler
protobuf==5.28.3           # Protocol buffers

# =============================================================================
# UNIVERSAL AI/ML STACK - Custom Summarization & Classification
# =============================================================================
torch==2.5.1               # PyTorch for fine-tuning
transformers==4.47.0       # Hugging Face transformers
sentence-transformers==3.3.1 # Embeddings and semantic similarity
datasets==3.1.0            # Dataset handling for fine-tuning
accelerate==1.1.1          # Training optimization

# =============================================================================
# VECTOR STORAGE & RAG - Memory and Retrieval
# =============================================================================
faiss-cpu==1.9.0           # Vector similarity search
numpy==2.2.0               # Numerical computing
scikit-learn==1.6.0        # ML utilities and metrics

# =============================================================================
# NATURAL LANGUAGE PROCESSING - Universal Extraction
# =============================================================================
spacy==3.8.2               # NLP pipeline for entity extraction
nltk==3.9.1                # Text processing utilities

# =============================================================================
# SECURITY & ENCRYPTION - Identra Vault Integration
# =============================================================================
cryptography==44.0.0       # Encryption/decryption for vault integration
pycryptodome==3.21.0        # Additional crypto utilities

# =============================================================================
# CONFIGURATION & UTILITIES
# =============================================================================
python-dotenv==1.0.1       # Environment configuration
pydantic-settings==2.6.1   # Settings management
python-multipart==0.0.18   # File upload support
httpx==0.28.1              # Async HTTP client
aiofiles==24.1.0           # Async file operations

# =============================================================================
# DEVELOPMENT & TESTING
# =============================================================================
pytest==8.3.4             # Testing framework
pytest-asyncio==0.24.0    # Async testing support
black==24.10.0             # Code formatting
isort==5.13.2              # Import sorting
mypy==1.13.0               # Type checking

# =============================================================================
# MONITORING & LOGGING
# =============================================================================
structlog==24.4.0          # Structured logging
prometheus-client==0.21.0  # Metrics collection
